requirements.txt
chromadb==0.5.0
sentence-transformers==2.6.0
gitpython==3.1.43
requests=2.31.0
pyyaml==6.0.1
fastapi==0.110.0
uvicorn==0.29.0